+++ CHANGELOG.md
@@ -55,6 +55,7 @@
   - Caching of Yarn dependencies is now handled by `actions/setup-node`
   - Migrated from `actions/create-release` and `actions/upload-release-asset` to `softprops/action-gh-release`
 - Fix Webpack only listening on IPv6 instead of all interfaces + update VS Code default styles
+- Added the `FieldList` and derivate `FieldConfigList` field types to the Settings UI
 
 ## v1.25.0 (2022-06-01)